+++ tests/testthat/tests.r
@@ -1,5 +1,0 @@
-context("Testing: Do nothing")
-
-test_that("whatever you like to test", {
-  expect_true( TRUE )
-})