+++ NEWS.md
@@ -1,6 +1,16 @@
 NEWS diffrprojects
 ==========================================================================
 
+
+version 0.1.0 // 2016-04-27 ... 
+--------------------------------------------------------------------------
+
+* FEATURES
+    - dp_text : an object for text (basic layout)
+    - text_read() : function for reading text
+    - text_snippet() : function for getting snippet of text
+
+
 version 0.0.1 // 2016-04-26 ... 
 --------------------------------------------------------------------------